OF-2535: Move channel-binding capability advertisement into SASLAuthentication
The appendChannelBindingCapabilityIfNeeded and isSaslAuthenticationFeature methods, along with their unit tests, are moved from LocalClientSession to SASLAuthentication. No behavior changes.

Relocating it here allows the incoming server session path to reuse the same implementation in a follow-up change.
